Fix Windows/clang-cl compiler warnings surfaced under -Weverything
- cl_types.h: frametime_index/frametime_count were uint8_t but
  FRAMETIME_COUNTER_SAMPLE_COUNT is 1024 -- the ring buffer's top 768
  entries were dead and the counter silently wrapped at 256. Widened
  to uint16_t.
- r_draw_2d.c: replaced a goto that jumped over a variable
  initialization with an early return duplicating the (two-line)
  reset.
- cm_entity.c: null_entity given an explicit `= { 0 }` (a const
  tentative definition with no initializer is invalid in C++, which
  clang-cl checks for even in C mode).
- cm_trace.c: Cm_EntityBounds restructured to a single return point.
- cmd.c, g_item.c: deleted dead code unreachable after a noreturn
  Com_Error/gi.Error call.
- common.h: Com_Error_f marked noreturn, matching Com_Shutdown --
  its body is a single noreturn Com_Error call.
- g_main.c: removed a duplicated `cvar_t *g_balance_shotgun_pellets;`
  declaration (copy-paste artifact).
- net_sock.c, master/main.c: explicit casts for SOCKET (64-bit on
  Windows)/size_t narrowing to the codebase's int32_t socket/length
  types, a (const char *) cast for a byte* passed as socket data, and
  an #if defined(_WIN32) split for an FD_SET sign-compare (SOCKET
  doesn't exist on POSIX).
- sv_client.c, cg_temp_entity.c: braced switch cases that declared
  locals without a block, which made jumping to a later case label
  (via the switch dispatch) skip their initialization -- invalid in
  C++, which clang-cl's -Weverything flags even in C mode.

Verified with `make -j8` (exit 0, no new warnings) locally; the
C++-compat-style warnings (jump-misses-init, default-const-init-var,
tentative-definition-compat) only surface under clang-cl's Windows
build and won't reproduce in a plain C compile here.

+++ src/net/net_sock.c
@@ -191,7 +191,7 @@ int32_t Net_Socket(net_addr_type_t type, const char *iface, in_port_t port) {
   switch (type) {
     case NA_BROADCAST:
     case NA_DATAGRAM:
-      if ((sock = socket(PF_INET, SOCK_DGRAM, IPPROTO_UDP)) == -1) {
+      if ((sock = (int32_t) socket(PF_INET, SOCK_DGRAM, IPPROTO_UDP)) == -1) {
         Com_Error(ERROR_DROP, "socket: %s\n", Net_GetErrorString());
       }
 
@@ -203,7 +203,7 @@ int32_t Net_Socket(net_addr_type_t type, const char *iface, in_port_t port) {
       break;
 
     case NA_STREAM:
-      if ((sock = socket(PF_INET, SOCK_STREAM, 0)) == -1) {
+      if ((sock = (int32_t) socket(PF_INET, SOCK_STREAM, 0)) == -1) {
         Com_Error(ERROR_DROP, "socket: %s\n", Net_GetErrorString());
       }
 
@@ -242,7 +242,7 @@ int32_t Net_Socket(net_addr_type_t type, const char *iface, in_port_t port) {
 int32_t Net_SocketListen(const char *iface, in_port_t port, int32_t backlog) {
   int32_t opt = 1;
 
-  const int32_t sock = socket(PF_INET, SOCK_STREAM, 0);
+  const int32_t sock = (int32_t) socket(PF_INET, SOCK_STREAM, 0);
   if (sock == -1) {
     Com_Warn("socket: %s\n", Net_GetErrorString());
     return -1;
@@ -289,7 +289,7 @@ int32_t Net_Accept(int32_t sock, net_addr_t *from) {
   net_sockaddr addr;
   socklen_t addr_len = sizeof(addr);
 
-  const int32_t client = accept(sock, (struct sockaddr *) &addr, &addr_len);
+  const int32_t client = (int32_t) accept(sock, (struct sockaddr *) &addr, &addr_len);
   if (client == -1) {
     return -1;
   }
@@ -310,7 +310,7 @@ int32_t Net_Accept(int32_t sock, net_addr_t *from) {
  * @return Bytes sent, or -1 on error.
  */
 ssize_t Net_Send(int32_t sock, const void *data, size_t len) {
-  return send(sock, data, len, 0);
+  return send(sock, data, (int32_t) len, 0);
 }
 
 /**
@@ -318,7 +318,7 @@ ssize_t Net_Send(int32_t sock, const void *data, size_t len) {
  * @return Bytes received, 0 on close, or -1 on error.
  */
 ssize_t Net_Recv(int32_t sock, void *data, size_t len) {
-  return recv(sock, data, len, 0);
+  return recv(sock, data, (int32_t) len, 0);
 }
 
 /**
